Run cleanup at start and all 20 mins
